feat: updated flag and divide form create and handler
