+++ commands/users/create-user.js
@@ -1,4 +1,10 @@
-import { SlashCommandBuilder, ActionRowBuilder, ButtonStyle, ButtonBuilder } from "discord.js";
+import {
+  SlashCommandBuilder,
+  ActionRowBuilder,
+  ButtonStyle,
+  ButtonBuilder,
+  MessageFlags,
+} from "discord.js";
 
 export default {
   data: new SlashCommandBuilder()
@@ -6,13 +12,12 @@ export default {
     .setDescription("Create new user in database"),
 
   async execute(interaction) {
-    console.log("creating form");
-    await interaction.deferReply({ ephemeral: true }); // acknowledges immediately
+    await interaction.deferReply({ flags: MessageFlags.Ephemeral }); // acknowledges immediately
     try {
       const { userService } = interaction.services;
       const isUser = await userService.getOneDiscordId(interaction.user.id);
       if (isUser) {
-        return interaction.editReply({ content: "User already exists.", ephemeral: true });
+        return interaction.editReply({ content: "User already exists." });
       }
 
       const button = new ButtonBuilder()
@@ -25,11 +30,11 @@ export default {
       return interaction.editReply({
         content: "You’re not in the database yet. Click below to create a profile.",
         components: [row],
-        ephemeral: true,
+        flags: MessageFlags.Ephemeral,
       });
     } catch (err) {
       console.log(err);
-      await interaction.editReply({ content: "Oops and error occurred", ephemeral: true });
+      await interaction.editReply({ content: "Oops and error occurred" });
     }
   },
 };
